spade_improved/requirements.txt
# SPADE 2.0 Requirements

# Core dependencies
numpy>=1.19.0
pillow>=8.0.0
matplotlib>=3.3.0

# Optional dependencies
# pytest>=6.0  # For testing
# pytest-cov>=2.10  # For coverage
